requirements.txt
alembic==1.9.4
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Email==1.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
